feat: restructure project by moving activity_logger and screenshot modules to src directory and update import paths
